Document live-pod durability and share turn queue
