+++ .github/workflows/conversation-contract.yml
@@ -10,6 +10,7 @@ on:
       - 'frontend/src/*.test.ts'
       - 'agent-runner/src/**'
       - 'codex-runner/src/**'
+      - 'runner-shared/**'
       - 'scripts/check-tank-conversation-contract.mjs'
       - 'docs/tank-conversation-protocol.md'
       - '.github/workflows/conversation-contract.yml'
@@ -22,6 +23,7 @@ on:
       - 'frontend/src/*.test.ts'
       - 'agent-runner/src/**'
       - 'codex-runner/src/**'
+      - 'runner-shared/**'
       - 'scripts/check-tank-conversation-contract.mjs'
       - 'docs/tank-conversation-protocol.md'
       - '.github/workflows/conversation-contract.yml'
