+++ AGENTS.md
@@ -6,7 +6,7 @@ Read `CLAUDE.md` for the project architecture and operational notes.
 
 Agent pods are not expected to have Docker. Do not report missing local Docker
 as a blocker. Run available repo checks first, then use PR CI as the normal
-container build gate: `.github/workflows/docker-build-check.yml` performs
+container build gate: `.github/workflows/docker-build-check.yaml` performs
 throwaway builds for every repo-owned image with `push: false`. If
 image-packaging feedback is needed before a PR is ready, manually dispatch that
 workflow with `git_ref`. Release/deploy workflows are the only path that
