+++ CLAUDE.md
@@ -21,7 +21,7 @@ The orchestrator was a Python FastAPI + kubernetes-asyncio app until 2026-05-11 
 Session pods intentionally do not ship Docker or a container runtime. Do not
 report missing local Docker as a blocker. Run available repo checks first
 (`go test`, `npm test`, `helm template`, etc.). The normal container
-build gate is PR CI: `.github/workflows/docker-build-check.yml` performs
+build gate is PR CI: `.github/workflows/docker-build-check.yaml` performs
 throwaway Docker builds for every repo-owned image with `push: false`. If an
 image-packaging change needs feedback before a PR is ready, manually dispatch
 that workflow with `git_ref`. Release/deploy workflows are the only path that
@@ -58,7 +58,7 @@ into a monolith.
 
 Headless Claude/Codex runs stream through the run WebSocket and Kubernetes exec helpers in `backend-go/internal/kubeexec/`. The session image bootstrap seeds agent state to skip onboarding prompts, exports the MCP bearer token from the projected SA token, and performs mode-aware credential setup. Claude, Codex, and Pi sessions use separate SHA-pinned images (`session.image`, `session.codexImage`, and `session.piImage`) built from the same Dockerfile with agent-specific CLIs and support packages baked in, so startup does not fetch skills/extensions at runtime.
 
-GUI sessions use long-lived pod-side SDK runners: `agent-runner` for Claude GUI and `codex-runner` for Codex GUI. The SPA submits user turns to `POST /api/sessions/{session_id}/turns`; the backend writes a durable Cosmos `turn-queue` row keyed by `client_nonce`, and the runner claims due rows before feeding the provider SDK. Claimed rows carry `claim_id`, `claimed_by`, `claim_expires_at`, and `attempt_count`; stale claims are reclaimed inside the same still-live session pod, and terminal queue updates require the current claim. `/agent-ws` is now the live transport only, with `/timeline` as the durable replay source. This queue is scoped to browser disconnects, orchestrator rollouts, and runner-process restarts inside an otherwise live session pod. If the session pod is deleted or dies, the session is dead and its `emptyDir` workspace is gone; do not describe session-pod resurrection as a durability goal. CLI/config modes still use the legacy `/run` stream-json path. Claude `ScheduleWakeup` turns are durable delayed `turn-queue` rows with `source=schedule-wakeup` and `available_at`.
+GUI sessions use long-lived pod-side SDK runners: `agent-runner` for Claude GUI and `codex-runner` for Codex GUI. The SPA submits user turns to `POST /api/sessions/{session_id}/turns`; the backend writes a durable Cosmos `turn-queue` row keyed by `client_nonce`, and the runner claims due rows before feeding the provider SDK. Claimed rows carry `claim_id`, `claimed_by`, `claim_expires_at`, and `attempt_count`; stale claims are reclaimed inside the same still-live session pod, and terminal queue updates require the current claim. `/agent-ws` is now the live transport only, with `/timeline` as the durable replay source. This queue is scoped to browser disconnects, orchestrator rollouts, and runner-process restarts inside an otherwise live session pod. Session-pod deletion or death is a terminal session lifecycle event and an explicit non-goal for messaging durability: the session is dead, its `emptyDir` workspace is gone, and Tank does not try to resurrect it. Do not describe session-pod resurrection as a durability goal or product gap unless the session lifecycle goal changes. CLI/config modes still use the legacy `/run` stream-json path. Claude `ScheduleWakeup` turns are durable delayed `turn-queue` rows with `source=schedule-wakeup` and `available_at`.
 
 The Codex path stays on per-turn CLI invocation — there is no Codex equivalent of the agent SDK to wrap, so the two dispatch shapes live next to each other in the orchestrator. See `backend-go/internal/sessions/dispatch.go`.
